Post event updates and bug fixes
