+++ data/pastMembers.tsx
@@ -1,0 +1,368 @@
+
+import { TeamMember } from "@/types";
+import {
+    Mail,
+    Linkedin,
+    Github,
+    Globe,
+} from "lucide-react";
+
+export const pastMembers: TeamMember[] = [
+    {
+        type: "alumni_faculty",
+        name: "Reshma Roychoudhuri",
+        roleAtACMHITK: "Faculty Sponsor(2013-2014)",
+        roleAtHITK:
+            "Assistant Professor, Department of Computer Science and Engineering, HITK",
+        currentRole:
+            "Assistant Professor, Department of Computer Science and Engineering, HITK",
+        image: "/Reshma-Roychoudhuri.jpeg",
+        bio: `With 6 plus years of experience in IT industry Reshma Roychoudhuri joined Heritage in 2012 to explore her passion  for teaching. She has always been involved in various club activities and responsibilities in school, college, and even in her IT workplace. Being entrusted with ACM started an enriching journey here in Heritage, one that she still cherishes. We ( she and the team then) started the annual flagship event, Brainstorm, in a bid to revive the club registrations. It is heartwarming to see the legacy continue and to witness the growth of the club under different faculty and student members`,
+        achievements: `She started the annual flagship event Brainstorm at ACM HITK Student Chapter`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:reshma.roychoudhuri@heritageit.edu",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/reshma-roychoudhuri-b660432b/",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_faculty",
+        name: "Dr. Dinabandhu Bhandari",
+        roleAtACMHITK: "Faculty Sponsor(2015-2016)",
+        roleAtHITK: `Professor in the Department of Computer Science and
+  Engineering (July 2013 - Present). Previously Associate
+  Professor (January 2013 - June 2017).`,
+        currentRole:
+            "Professor, Department of Computer Science and Engineering, HITK",
+        image: "/Dr-Dinabandhu-Bhandari.jpeg",
+        bio: `Dr. Dinabandhu Bhandari has accumulated over 30 years of experience in academia and industry. His professional journey includes tenures at Heritage Institute of Technology, Indian Statistical Institute, TCS, Capgemini, and Cognizant. He specializes in Soft Computing, Pattern Recognition, and Machine Learning. Dr. Bhandari holds a PhD and M.Tech in Computer Science from Jadavpur University.`,
+        achievements: `Served as a technical consultant implementing intelligent systems. Developed innovative methodologies for article/book structuring and mining sector solutions. Authored research papers with over 1,600 Google Scholar citations. Ranked #1 in citations at HITK by AD Scientific Index.`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:dinabandhu.bhandari@heritageit.edu",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://in.linkedin.com/in/dinabandhu-bhandari-22b68111",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Website",
+                url: "https://www.researchgate.net/profile/Dinabandhu-Bhandari",
+                icon: <Globe className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_faculty",
+        name: "Dr. Partha Basuchowdhuri",
+        roleAtACMHITK: "Faculty Sponsor(2016-2018)",
+        roleAtHITK: `Assistant Professor (June 2010 - June 2018), Associate
+  Professor (July 2018 - June 2019) in the Dept. of Computer
+  Science and Engineering. Founding faculty sponsor of the
+  ACM student chapter at HITK.`,
+        currentRole:
+            "Associate Professor, School Chair @ School of Mathematical and Computational Sciences, Indian Association for the Cultivation of Science, Kolkata",
+        image: "/Dr-Partha-Basuchowdhuri.png",
+        bio: `Dr. Partha Basuchowdhuri holds Bachelor's, Master's, and PhD degrees from IIEST Shibpur, Louisiana State University, and Jadavpur University, respectively. His research at HITK focused on social network analysis and data mining. He completed post-doctoral tenures at Queen's University Belfast and the Digital Enterprise Research Institute (DERI) in Ireland. Currently, he serves as an Associate Professor and School Chair at the Indian Association for the Cultivation of Science (IACS).`,
+        achievements: `Led the DBT-VInnova project titled "AI-based Detection of Acute Respiratory Distress Syndrome (AI-DARDS)" with a 25 Lakh grant for IACS. Received the Best Teaching Award at Heritage Institute of Technology in 2015. Published in PAKDD, ASONAM, CoDS-COMAD, KAIS, IEEE Transactions on Big Data, and Innovations in Systems and Software Engineering. Awarded US National Science Foundation funding for Master's thesis research.`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:partha.basuchowdhuri@iacs.res.in",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://in.linkedin.com/in/partha-basuchowdhuri-44043a126",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Website",
+                url: "https://www.researchgate.net/profile/Partha-Basuchowdhuri",
+                icon: <Globe className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_faculty",
+        name: "Dr. Sabyasachee Banerjee",
+        roleAtACMHITK: "Faculty Sponsor(2018-2025)",
+        roleAtHITK:
+            "Assistant Professor, Department of Computer Science and Engineering",
+        currentRole:
+            "Assistant Professor, Department of Computer Science and Engineering, HITK",
+        image: "/Dr-Sabyasachi-Banerjee.jpg",
+        bio: `Sabyasachee Banerjee (Dr.) is an Assistant Professor in the Department of Computer Science and Engineering at Heritage Institute of Technology, Kolkata. He holds a Ph.D. in Computer Science & Engineering from Jadavpur University, with research focused on layout optimization, thermal issues, and testing of 3D Integrated Circuits. With over a decade of teaching experience, he has guided multiple undergraduate and postgraduate projects, resulting in several international publications. Dr. Banerjee has served as a reviewer for reputed journals, coordinated academic events as the faculty sponsor of the ACM Student Chapter at HITK. His research interests include 3D SoC design and testing, VLSI design automation, and optimization algorithms.`,
+        achievements: null,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:sabyasachi.banerjee@heritageit.edu",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/dr-sabyasachee-banerjee-13578319",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Website",
+                url: "https://scholar.google.co.in/citations?user=hH3VlT0AAAAJ&hl=en",
+                icon: <Globe className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Ashmi Banerjee",
+        roleAtACMHITK: "Webmaster(2014-15), Chair (2015-16)",
+        roleAtHITK: "CSE'16",
+        currentRole: "PhD Student at Technical University Munich",
+        image: "/Ashmi-Banerjee.jpg",
+        bio: "CSE 2016 batch, MSc TU Munich 2019, ongoing PhD at TUM",
+        achievements: `https://ashmibanerjee.com/#awards`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:ashmi.banerjee@tum.de",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/ashmi-banerjee/",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Website",
+                url: "https://ashmibanerjee.com",
+                icon: <Globe className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Ananya Majumdar",
+        roleAtACMHITK: "Membership Head(2021-2022), Vice Chair(2022-2023)",
+        roleAtHITK: "CSE'23",
+        currentRole: "Technology Consulting - Analyst I, EY",
+        image: "/Ananya-Majumdar.png",
+        bio: "Ananya Majumdar graduated from Heritage Institute of Technology in 2023, thereafter, started working at Ernst & Young.",
+        achievements: null,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:majumdarananya2000@gmail.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "http://linkedin.com/in/ananya-majumdar-09712b254",
+                icon: <Linkedin className="h-5 w-5" />,
+            }
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Agnish Ghosh",
+        roleAtACMHITK: "Webmaster(2022-2023)",
+        roleAtHITK: "CSE'24",
+        currentRole:
+            "Contributor/Maintainer at the Ethereum Foundation and Ethereum Core Developer at Status",
+        image: "/Agnish-Ghosh.png",
+        bio: `Agnish Ghosh is a highly accomplished individual, distinguished by his profound expertise and dedication to advancing the Ethereum ecosystem. He possesses extensive experience in cutting-edge blockchain development, intricate cryptography, and robust decentralized systems. His comprehensive skill set ranges from optimizing cryptographic primitives vital for Ethereum upgrades to spearheading the implementation of critical protocol improvements. He graduated from Heritage Institute of Technology, Kolkata with a B.Tech in CSE in 2024.`,
+        achievements: `As a Contributor/Maintainer on the c-kzg-4844 cryptography team at the Ethereum Foundation, he developed computationally fast and low-memory overhead cryptography for verifying blobs and data columns, crucial for Ethereum client teams and ensuring backward compatibility for Deneb and Peerdas upgrades.
+          As an Ethereum Core Developer at Status, he is instrumental in achieving statelessness by migrating Ethereum from Merkle trees to Verkle trees, and is the sole project owner and programmer of Nimbus Ethereum CL’s Peerdas implementation, a key upgrade for optimizing blob transactions expected with the Fusaka hard fork. He contributed to Rollups as a Service and explored secure privacy-preserving groups on EVM chains as a Software Engineer - Blockchain & Cryptography at BaseLayer. At Cipherem, he focused on EVM Equivalent zkRollups and finding solutions for Ethereum infrastructure as a Research And Development Engineer, and was involved in building a self-sovereign social network as a Blockchain Developer. He completed a 5-month internship as a Protocol Fellow at Ethereum and was a Devconnect Scholarship Recipient. He co-founded and led R&D at Metafy. He served as Webmaster for the ACM chapter at HITK and held the position of Technical Lead at Google Developer Students'​ Club HIT. As an Undergraduate Student Researcher at the Indian Association for the Cultivation of Science, his research included prompt generation for Zero-Shot Relation Extraction and Contrastive Learning in Distantly Supervised Named Entity Relation Extraction, applying NLP principles like Coreferencing and developing optimized O(n) clustering algorithms that reduced time in establishing inter-sentence relations by 96%. He also worked as a Campus Builder for Binance. He co-founded and developed the backend for an app aimed at verifying and providing COVID-19 resources as a Co-Founder and Backend Developer at Kolkata Covid Resources. As an SDE Intern at Coinhunt.cc, he worked on a randomized rendering algorithm in React.js, improving rendering latencies by 80%, and built a crypto-payments gateway by integrating APIs from Metamask, Phantom Wallet, and Coinbase, optimizing user payment experience by 75%.`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:iamagnix@gmail.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/agnish-ghosh-15902a211/",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "GitHub",
+                url: "https://github.com/agnxsh",
+                icon: <Github className="h-5 w-5" />,
+            },
+            {
+                name: "Website",
+                url: "https://www.agnish.in/",
+                icon: <Globe className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Protyasha Kundu",
+        roleAtACMHITK: "Chair & Treasurer(2023-2025)",
+        roleAtHITK: "CSE'24",
+        currentRole: "MTech Student at ISI Kolkata",
+        image: "/Protyasha-Kundu.jpg",
+        bio: `I am Protyasha Kundu. I am currently pursuing MTech in Cryptology and Security at ISI Kolkata. I have major interest in Networks and Network Security System. From July 2024 till March 2025 i was working for Ernst & Young as a Senior Analyst in Tech Consulting for GPS service line. I am a 2024 graduate from Heritage Institute of Technologym, Kolkata.`,
+        achievements: `She was honored with the Best Paper Award at ICISS 2024. Academically, she performed exceptionally well, ranking in the Top 11 in the TIFR GS exam in 2025 and securing the 3rd position in the CSE 2020 Batch at HITK.For ACM HITK, she has been actively involved in organizing and hosting numerous successful events. These include multiple guest lectures featuring prominent speakers such as Chandra Shekhar Mukherjee, Asmita Banerjee, and Prof. Sourav Chakraborty. Additionally, she played a key role in organizing the TechQuiz under the ACM HITK Student Chapter.`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:protyashakundu12901@gmail.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/mwlite/profile/in/protyasha-kundu-69744a214/",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Swarnadeep Saha",
+        roleAtACMHITK: "Secretary(2023-24). Treasurer(2022-23)",
+        roleAtHITK: "CSE'24",
+        currentRole: "Mtech(CSE) Student at IIIT Hyderabad",
+        image: "/Swarnadeep-Saha.jpg",
+        bio: "Deeply curious masters student, looking for novel problems to learn and solve. Interested in solving practical problems with impact on real people. Currently pursuing M.Tech in IIIT Hyderabad (CSE)",
+        achievements: `1. Recipient of Reliance Foundation Scholarship awarded to top 100 PG students across the country.
+  2. Received Best Paper Award in International Conference for publishing novel research.
+  3. PGEE AIR 5.`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:swarnadeep0602@gmail.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/spotswan/",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Github",
+                url: "https://github.com/spotswan",
+                icon: <Github className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Rajarshi Saha",
+        roleAtACMHITK: "Asst. Membership Head(2023-2024), Membership Head(2024-25)",
+        roleAtHITK: "CSE'25",
+        currentRole: "Associate Software Engineer, EY GDS",
+        image: "/Rajarshi-Saha.jpeg",
+        bio: "Rajarshi Saha is currently working as an Associate Software Engineer at EY GDS. He is a full-stack developer with experience across diverse technology domains, including the development of Progressive Web Apps and building multiple applications that involve Blockchain and Machine Learning. From 2023 to 2025, he actively organized several tech talks and events for the ACM HITK Student Chapter.",
+        achievements: `Rajarshi organized Brainstorm'24, multiple tech-talks. He represented ACM-HITK Chapter at ACM Regional Meet 2024 and was actively involved in driving student membership registrations for the 2023-25 term.`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:saharajarshi14@gmail.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/rajarshi-saha14/",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Github",
+                url: "https://github.com/rajarshigit14",
+                icon: <Github className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Swapnil Aichbhaumik",
+        roleAtACMHITK: "Board Member, Joint Secretary, Treasurer(2022-2025)",
+        roleAtHITK: "CSE'25",
+        currentRole: "Associate Software Engineer at EY GDS",
+        image: "/Swapnil-Aichbhaumik.jpg",
+        bio: `Swapnil Aichbhaumik is a Computer Science and Engineering graduate with a primary focus on Product Engineering and Development. He gained practical experience as a Product Engineering Intern at SETV Global and is currently serving as an Associate Software Engineer at EY GDS.`,
+        achievements: `For ACM HITK, he organized Brainstorm '23 and Tech quiz' 23`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:Swapnil.Aichbhaumik@gds.ey.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: `http://www.linkedin.com/in/swapnil-aichbhaumik-28a980224`,
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Github",
+                url: "https://github.com/Swap-nil-2003",
+                icon: <Github className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Upasana Roy",
+        roleAtACMHITK: "Chairperson(2024-25) , Vice-Chairperson(2023-24)",
+        roleAtHITK: "CSE'25",
+        currentRole: "SDE 1, Cozeva - Applied Research Works India Private Limited",
+        image: "/Upasana-Roy.jpg",
+        bio: `I am currently working as SDE 1 at Cozeva - Applied Research Works India Private Limited. At ACM, I have collaborated with other board members to organize different events like Tech quizzes, Tech Talks, Seminars, Brainstorming events for better Chapter engagement among the students and attended ACM Regional Chapter Meet to represent our HITK ACM chapter.`,
+        achievements:
+            "Currently working as SDE 1 at Cozeva - Applied Research Works India Private Limited (from June 2025). There I have contributed to Full stack Projects.Organized several events like Brainstorm'24, Brainstorm'25, etc. Organized Tech Talks about Birthday Paradoxes, God Forever Geometrizes, Unlocking the secret of language with large language models, etc. Represented our HITK chapter at ACM Regional meet 2024. Also presented our chapter to the freshers of our college for increasing engagement.",
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:upasanaroy2002@gmail.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/upasana-roy-27874a229",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Github",
+                url: "https://github.com/Upasana-Roy",
+                icon: <Github className="h-5 w-5" />,
+            },
+        ],
+    },
+    {
+        type: "alumni_student",
+        name: "Devjit Sarkar",
+        roleAtACMHITK: "Joint Secretary(2024-2025)",
+        roleAtHITK: "CSE'26",
+        currentRole: "Summer Intern, Hyland",
+        image: "/Devjit-Sarkar.jpg",
+        bio: "I am a 4th-year Computer Science student currently preparing for my Master's while actively seeking opportunities in the tech industry. As the Joint Secretary of ACM HITK, I contributed by designing event flyers and overseeing the successful execution of Brainstorm 2025. My involvement has helped strengthen the chapter’s outreach and event planning efforts.",
+        achievements: `I have achieved four 10.0 SGPAs over six semesters, reflecting my consistent academic performance. I'm currently contributing to a JNI-based reimplementation of the Phoenix OCR engine for text and table extraction—an effort I'm particularly proud of under my internship at Hyland, Kolkata. Additionally, I was part of the winning team at the National Hackathon StatusCode1, hosted by IIIT-Kalyani at IISER-Kolkata, a testament to our teamwork and technical innovation. Winner of ACM's Flagship event BrainStorm in 2023, and 2024`,
+        socials: [
+            {
+                name: "Email",
+                url: "mailto:sardevjit7203@gmail.com",
+                icon: <Mail className="h-5 w-5" />,
+            },
+            {
+                name: "LinkedIn",
+                url: "https://www.linkedin.com/in/devjit-sarkar-5185b123b/",
+                icon: <Linkedin className="h-5 w-5" />,
+            },
+            {
+                name: "Github",
+                url: "https://github.com/devjit-sarkar",
+                icon: <Github className="h-5 w-5" />,
+            },
+        ],
+    },
+];
