+++ types/index.ts
@@ -19,3 +19,23 @@ export interface TeamMember {
     achievements: string | null;
     socials: SocialLink[];
 }
+
+export interface Event {
+    title: string;
+    date: string;
+    time?: string;
+    location: string;
+    type: string;
+    cardDescription: string;
+    description?: string;
+    status?: string;
+    registration_deadline?: string;
+    registration_link?: string;
+    bannerImage?: string;
+    speaker?: string;
+    capacity?: string;
+    attendees?: number;
+    photos?: string[];
+    image?: string;
+    pictures?: string[];
+}
